requirements.txt
astroid==2.15.5
blinker==1.6.2
blosc2==2.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
coverage==7.2.7
cycler==0.11.0
Cython==0.29.35
decorator==5.1.1
dill==0.3.6
Flask==2.3.2
fonttools==4.40.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
msgpack==1.0.5
numexpr==2.8.4
numpy==1.24.3
obspy==1.4.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
platformdirs==3.5.3
pluggy==1.0.0
psutil==5.9.5
psycopg2==2.9.6
py-cpuinfo==9.0.0
pylint==2.17.4
pyparsing==3.0.9
pytest==7.3.2
pytest-cov==4.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
scipy==1.10.1
-e git+https://github.com/rizac/sdaas.git@6647d5dd3e802019a5fcdaf9758f912242a105c9#egg=sdaas
six==1.16.0
SQLAlchemy==2.0.16
-e git+https://github.com/rizac/stream2segment.git@129cefaf2da0dac8d9bf22d2e44f0f5c8989d8a6#egg=stream2segment
tables==3.8.0
tomlkit==0.11.8
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
